DABOM-229 refactor: 알림 이벤트 관련 인터페이스 및 구현체 디렉토리 이동
