+++ src/main/java/com/project/domain/usage/infra/messaging/NotificationEventPublisher.java
@@ -1,4 +1,4 @@
-package com.project.domain.notification.service.port;
+package com.project.domain.usage.infra.messaging;
 
 import com.project.global.event.dto.notification.NotificationPayload;
